Properly handle editor editable state changes in ResizableImageComponent
